+++ src/main/java/me/navidrohim/jmws/JMWS.java
@@ -20,7 +20,7 @@ public class JMWS
 {
     public static final String MODID = "jmws";
     public static final String NAME = "JourneyMap Waypoint Syncing";
-    public static final String VERSION = "1.1.7-1.12.2-beta.3";
+    public static final String VERSION = "1.1.7-1.12.2";
 
     public static Logger logger;